+++ packages/wix-ui-core/src/components/image/image.uni.driver.ts
@@ -5,19 +5,14 @@ import {
 } from 'wix-ui-test-utils/base-driver';
 import { StylableDOMUtil } from '@stylable/dom-test-kit';
 import style from './image.st.css';
-import {
-  imageDriverFactory as publicImageDriverFactory,
-  ImageDriver as ImagePublicDriver,
-} from './image.uni.driver';
 
-export interface ImageDriver extends ImagePublicDriver {
-  getResizeMode(): Promise<string>;
-  getSrcSet(): Promise<string>;
+export interface ImageDriver extends BaseUniDriver {
+  getSrc(): Promise<string | null>;
+  getAlt(): Promise<string>;
+  getLoadStatus(): Promise<string>;
 }
 
 export const imageDriverFactory = (base: UniDriver): ImageDriver => {
-  const publicDriver = publicImageDriverFactory(base);
-
   const getStyleState = async (styleState: string) => {
     const el = (await base.getNative()) as Element;
     const domUtils = new StylableDOMUtil(style, el);
@@ -25,8 +20,9 @@ export const imageDriverFactory = (base: UniDriver): ImageDriver => {
   };
 
   return {
-    ...publicDriver,
-    getResizeMode: async () => getStyleState('resizeMode'),
-    getSrcSet: () => base.attr('srcSet'),
+    ...baseUniDriverFactory(base),
+    getSrc: () => base.attr('src'),
+    getAlt: () => base.attr('alt'),
+    getLoadStatus: async () => getStyleState('loadState'),
   };
 };
