+++ packages/wix-ui-core/src/components/video/Video.uni.driver.ts
@@ -1,0 +1,18 @@
+import {
+  BaseUniDriver,
+  baseUniDriverFactory,
+} from 'wix-ui-test-utils/base-driver';
+import { UniDriver } from 'wix-ui-test-utils/unidriver';
+
+export interface IVideoDriver extends BaseUniDriver {
+  /** returns player name */
+  getPlayerName(): Promise<string>;
+}
+
+export const videoDriverFactory = (base: UniDriver): IVideoDriver => {
+  return {
+    ...baseUniDriverFactory(base),
+    getPlayerName: async () =>
+      base.$('[data-player-name]').attr('data-player-name'),
+  };
+};
