+++ packages/wix-ui-core/src/testkit/index.ts
@@ -3,33 +3,33 @@ import {
   uniTestkitFactoryCreator,
 } from 'wix-ui-test-utils/vanilla';
 
-import {inputDriverFactory} from '../components/input/Input.driver';
+import { inputDriverFactory } from '../components/input/Input.driver';
 export const inputTestkitFactory = testkitFactoryCreator(inputDriverFactory);
 
-import {paginationDriverFactory} from '../components/pagination/Pagination.driver';
+import { paginationDriverFactory } from '../components/pagination/Pagination.driver';
 export const paginationTestkitFactory = testkitFactoryCreator(
-  paginationDriverFactory
+  paginationDriverFactory,
 );
 
-import {tooltipDriverFactory} from '../components/tooltip/Tooltip.driver';
+import { tooltipDriverFactory } from '../components/tooltip/Tooltip.driver';
 export const tooltipTestkitFactory = testkitFactoryCreator(
-  tooltipDriverFactory
+  tooltipDriverFactory,
 );
 
-import {dividerDriverFactory} from '../components/deprecated/divider/Divider.driver';
+import { dividerDriverFactory } from '../components/deprecated/divider/Divider.driver';
 export const dividerTestkitFactory = testkitFactoryCreator(
-  dividerDriverFactory
+  dividerDriverFactory,
 );
 
-import {checkboxDriverFactory} from '../components/checkbox/Checkbox.driver';
+import { checkboxDriverFactory } from '../components/checkbox/Checkbox.driver';
 export const checkboxTestkitFactory = testkitFactoryCreator(
-  checkboxDriverFactory
+  checkboxDriverFactory,
 );
 
 //Stylable
-import {toggleSwitchDriverFactory} from '../components/toggle-switch/ToggleSwitch.driver';
+import { toggleSwitchDriverFactory } from '../components/toggle-switch/ToggleSwitch.driver';
 export const toggleSwitchTestkitFactory = testkitFactoryCreator(
-  toggleSwitchDriverFactory
+  toggleSwitchDriverFactory,
 );
 
 import {
@@ -37,18 +37,18 @@ import {
   ButtonDriver,
 } from '../components/deprecated/button/Button.driver';
 export const buttonTestkitFactory = testkitFactoryCreator<ButtonDriver>(
-  buttonDriverFactory
+  buttonDriverFactory,
 );
-export {ButtonDriver};
+export { ButtonDriver };
 
 import {
   buttonNextDriverFactory,
   ButtonNextDriver,
-} from '../components/button-next/button-next.driver';
+} from '../components/button-next/button-next.uni.driver';
 export const buttonNextTestkitFactory = uniTestkitFactoryCreator<
   ButtonNextDriver
 >(buttonNextDriverFactory);
-export {ButtonNextDriver};
+export { ButtonNextDriver };
 
 import {
   linearProgressBarDriverFactory,
@@ -57,7 +57,7 @@ import {
 export const linearProgressBarTestkitFactory = testkitFactoryCreator<
   LinearProgressBarDriver
 >(linearProgressBarDriverFactory);
-export {LinearProgressBarDriver};
+export { LinearProgressBarDriver };
 
 import {
   circularProgressBarDriverFactory,
@@ -66,7 +66,7 @@ import {
 export const circularProgressBarTestkitFactory = testkitFactoryCreator<
   CircularProgressBarDriver
 >(circularProgressBarDriverFactory);
-export {CircularProgressBarDriver};
+export { CircularProgressBarDriver };
 
 import {
   badgeDriverFactory as stylableBadgeDriverFactory,
@@ -75,37 +75,37 @@ import {
 export const stylableBadgeTestkitFactory = testkitFactoryCreator<
   StylableBadgeDriver
 >(stylableBadgeDriverFactory);
-export {StylableBadgeDriver};
+export { StylableBadgeDriver };
 
-import {radioButtonDriverFactory} from '../components/radio-button/RadioButton.driver';
+import { radioButtonDriverFactory } from '../components/radio-button/RadioButton.driver';
 export const radioButtonTestkitFactory = testkitFactoryCreator(
-  radioButtonDriverFactory
+  radioButtonDriverFactory,
 );
 
-import {autocompleteDriverFactory} from '../components/autocomplete/Autocomplete.driver';
+import { autocompleteDriverFactory } from '../components/autocomplete/Autocomplete.driver';
 export const autocompleteTestkitFactory = testkitFactoryCreator(
-  autocompleteDriverFactory
+  autocompleteDriverFactory,
 );
 
-import {sliderDriverFactory} from '../components/slider/Slider.driver';
+import { sliderDriverFactory } from '../components/slider/Slider.driver';
 export const sliderTestkitFactory = testkitFactoryCreator(sliderDriverFactory);
 
-import {addressInputDriverFactory} from '../components/address-input/AddressInput.driver';
+import { addressInputDriverFactory } from '../components/address-input/AddressInput.driver';
 export const addressInputTestkitFactory = testkitFactoryCreator(
-  addressInputDriverFactory
+  addressInputDriverFactory,
 );
 
-import {labelDriverFactory} from '../components/deprecated/label/Label.driver';
+import { labelDriverFactory } from '../components/deprecated/label/Label.driver';
 export const labelTestkitFactory = testkitFactoryCreator(labelDriverFactory);
 
-import {timePickerDriverFactory} from '../components/time-picker/TimePicker.driver';
+import { timePickerDriverFactory } from '../components/time-picker/TimePicker.driver';
 export const timePickerTestkitFactory = testkitFactoryCreator(
-  timePickerDriverFactory
+  timePickerDriverFactory,
 );
 
-import {labelWithOptionsDriverFactory} from '../components/label-with-options/LabelWithOptions.driver';
+import { labelWithOptionsDriverFactory } from '../components/label-with-options/LabelWithOptions.driver';
 export const labelWithOptionsTestkitFactory = testkitFactoryCreator(
-  labelWithOptionsDriverFactory
+  labelWithOptionsDriverFactory,
 );
 
 import {
@@ -113,13 +113,13 @@ import {
   ThumbnailDriver,
 } from '../components/thumbnail/Thumbnail.driver';
 export const thumbnailTestkitFactory = testkitFactoryCreator<ThumbnailDriver>(
-  thumbnailDriverFactory
+  thumbnailDriverFactory,
 );
-export {ThumbnailDriver};
+export { ThumbnailDriver };
 
-import {popoverDriverFactory} from '../components/popover/Popover.driver';
+import { popoverDriverFactory } from '../components/popover/Popover.driver';
 export const popoverTestkitFactory = testkitFactoryCreator(
-  popoverDriverFactory
+  popoverDriverFactory,
 );
 
 import {
@@ -129,4 +129,4 @@ import {
 export const linearProgressBarUniTestkitFactory = uniTestkitFactoryCreator<
   LinearProgressBarUniDriver
 >(linearProgressBarUniDriverFactory);
-export {LinearProgressBarUniDriver};
+export { LinearProgressBarUniDriver };
