+++ packages/wix-ui-core/src/testkit/protractor.ts
@@ -24,7 +24,7 @@ export { ButtonDriver };
 import {
   buttonNextDriverFactory,
   ButtonNextDriver,
-} from '../components/button-next/button-next.driver';
+} from '../components/button-next/button-next.uni.driver';
 export const buttonNextTestkitFactory = protractorUniTestkitFactoryCreator(
   buttonNextDriverFactory,
 );
@@ -215,7 +215,7 @@ export { NavStepperDriver };
 import {
   captchaDriverFactory,
   CaptchaDriver,
-} from '../components/captcha/Captcha.driver';
+} from '../components/captcha/Captcha.uni.driver';
 export const captchaTestkitFactory = protractorUniTestkitFactoryCreator<
   CaptchaDriver
 >(captchaDriverFactory);
